% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_depth.R
\name{calculate_depth}
\alias{calculate_depth}
\title{Calculates the average sequencing depth of a bam/sam file.}
\usage{
calculate_depth(input, keep_csv = FALSE, path_db = NULL)
}
\arguments{
\item{input}{A sorted bam/sam file, (run_bwa already sorts its output)}

\item{keep_csv}{If wanted to keep csv file with each read depth.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
The average sequencing depth value of the input bam/sam file.
}
\description{
Calculates the average sequencing depth of a bam/sam file.
}
\examples{
\dontshow{if (Sys.which("samtools") != "") withAutoprint(\{ # examplesIf}
#1. Load .bam test file
temp_bam <- system.file("extdata", "test_mapped.bam", package = "AEkcompgen")
#2. Run caclulate depth function
calculate_depth(input =  temp_bam)
\dontshow{\}) # examplesIf}
}
