% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_1_seq_plot.R
\name{cluster_1_seq_plot}
\alias{cluster_1_seq_plot}
\title{Plots the quantity of clusters with 1 k-mer, and the kmer count}
\usage{
cluster_1_seq_plot(cluster_1_seq_df, marker_label = NULL)
}
\arguments{
\item{cluster_1_seq_df}{Dataframe from  the get_clusters_1_seq function.}

\item{marker_label}{A list with the names of the markers wished to be in the diagram, KEEP THE MARKERS' ORDER IN MIND}
}
\value{
A Venn Diagram based on the cluster distribuition among markers
}
\description{
Plots the quantity of clusters with 1 k-mer, and the kmer count
}
\examples{
# 1. Create temporary dataframe
cluster_1_seq_df <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A_1_69", "B_1_34", "C_1_35", "B_2_8", "C_2_9"),
  V3 = c("A", "B", "C", "B", "C"))

# 2. Run function with temporary file
plots_1_seq <- cluster_1_seq_plot(cluster_1_seq_df)
# 3. View result
print(plots_1_seq)
}
