% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sharing_relation.R
\name{cluster_sharing_relation}
\alias{cluster_sharing_relation}
\title{Joins different datasets labels based on the cluster they belong to, and returns them in list form to generate plots about the sharing relation}
\usage{
cluster_sharing_relation(dataframe_cd, path_db = NULL)
}
\arguments{
\item{dataframe_cd}{A dataframe from the run_cd_hit function.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
A list of lists containing the the clusters and in which the dataset is present
}
\description{
Joins different datasets labels based on the cluster they belong to, and returns them in list form to generate plots about the sharing relation
}
\examples{
# 1. Create temporary dataframe
 collapsed_datasets_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))
# 2. Run function with temporary file
cluster_sharing_list <- cluster_sharing_relation(collapsed_datasets_test)
# 3. View result
print(cluster_sharing_list)

}
