% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_labels.R
\name{collapse_labels}
\alias{collapse_labels}
\title{Joins different labels in a single line, based on the cluster they belong to.}
\usage{
collapse_labels(dataframe_cd, path_db = NULL)
}
\arguments{
\item{dataframe_cd}{A dataframe from the run_cd_hit function.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
A dataframe containing the collumns "cluster" and "labels".
}
\description{
Joins different labels in a single line, based on the cluster they belong to.
}
\examples{
# 1. Create temporary dataframe
cdhit_out <- data.frame(V1 = c(1, 1, 2, 3, 3),
 V2 = c("infantum_1_28", "major_1_33", "infantum_2_2", "infantum_3_19", "major_2_49"))

# 2. Run function with temporary file
collapsed_file <- collapse_labels(cdhit_out)
# 3. View result
print(collapsed_file)
}
