% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_euler_clusters.R
\name{dataset_euler_clusters}
\alias{dataset_euler_clusters}
\title{Creates an Euler Diagram based on the cluster distribuition among dataset
The suggested maximum number of dataset is 4.}
\usage{
dataset_euler_clusters(
  collapsed_dataset,
  dataset_label = NULL,
  color_pallete = "Set 2",
  title_name = "Cluster distribution across dataset"
)
}
\arguments{
\item{collapsed_dataset}{A dataframe from the collapse_datasets function}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE dataset' ORDER IN MIND}

\item{color_pallete}{A color pallete native from R, default is "Set 2"}

\item{title_name}{The title for the plot.}
}
\value{
A Euler diagram based on the cluster distribuition among datasets
}
\description{
Creates an Euler Diagram based on the cluster distribuition among dataset
The suggested maximum number of dataset is 4.
}
\examples{
# 1. Create temporary dataframe
collapsed_dataset_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))

# 2. Run function with temporary file
euler_cd <- dataset_euler_clusters(collapsed_dataset_test)
# 3. View result
print(euler_cd)
}
