% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_upset_clusters.R
\name{dataset_upset_clusters}
\alias{dataset_upset_clusters}
\title{Creates an UpSet Plot based on the cluster distribuition among datasets.
Recommeded whe working with a lot of datasets}
\usage{
dataset_upset_clusters(
  collapsed_datasets,
  dataset_label = NULL,
  color_pallete = "PuBuGn",
  empty_intersection = NULL,
  file_name = NULL
)
}
\arguments{
\item{collapsed_datasets}{A dataframe from the collapse_datasets function}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE datasetS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn".}

\item{empty_intersection}{If the empty intersectios are to be displayed put "on" or not, default is NULL}

\item{file_name}{The name for the file in case you want the plot to be saved automatically}
}
\value{
A Venn Diagram based on the cluster distribuition among datasets
}
\description{
Creates an UpSet Plot based on the cluster distribuition among datasets.
Recommeded whe working with a lot of datasets
}
\examples{
# 1. Create temporary dataframe
collapsed_datasets_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))

# 2. Run function with temporary file
upset_cd <- dataset_upset_clusters(collapsed_datasets_test)
# 3. View result
print(upset_cd)
}
