% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_upset_kmers.R
\name{dataset_upset_kmers}
\alias{dataset_upset_kmers}
\title{Creates an Upset Plot based on the kmer sequence distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more markes it c}
\usage{
dataset_upset_kmers(
  ...,
  path_db = NULL,
  dataset_label = NULL,
  color_pallete = "PuBuGn",
  empty_intersection = NULL,
  file_name = NULL
)
}
\arguments{
\item{...}{A list of at least two dataframes from containing renamed headers and sequences}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE datasetS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}

\item{empty_intersection}{If the empty intersectios are to be displayed put "on" or not, default is NULL}

\item{file_name}{The name for the file in case you want the plot to be saved automatically}
}
\value{
An Upset plot based on the kmer distribuition among datasets
}
\description{
Creates an Upset Plot based on the kmer sequence distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more markes it c
}
\examples{
# 1. Create temporary dataframes
data_1 <- data.frame(id = c("tag1_1_1", "tag1_2_2", "tag1_3_3", "tag1_4_4", "tag1_5_5"),
 seq = c("AAAG", "AGAG", "CTGC", "AAAA", "GAGA"))
data_2 <- data.frame(id = c("tag2_1_1", "tag2_2_2", "tag2_3_3", "tag2_4_4", "tag2_5_5"),
 seq = c("AATG", "AGAG", "CAGC", "AATA", "GAGA"))

# 2. Run function with temporary file
venn_k <- dataset_upset_kmers(data_1, data_2)
# 3. View result
print(venn_k)
}
