% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_venn_clusters.R
\name{dataset_venn_clusters}
\alias{dataset_venn_clusters}
\title{Creates a Venn Diagram based on the cluster distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more markes it c}
\usage{
dataset_venn_clusters(
  collapsed_datasets,
  dataset_label = NULL,
  color_pallete = "PuBuGn",
  plot_title = "Cluster distribution across datasets"
)
}
\arguments{
\item{collapsed_datasets}{A dataframe from the collapse_datasets function}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE datasetS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}

\item{plot_title}{The title for the plot.}
}
\value{
A Venn Diagram based on the cluster distribuition among datasets
}
\description{
Creates a Venn Diagram based on the cluster distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more markes it c
}
\examples{
# 1. Create temporary dataframe
collapsed_datasets_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))

# 2. Run function with temporary file
venn_cd <- dataset_venn_clusters(collapsed_datasets_test)
# 3. View result
print(venn_cd)
}
