% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_line.R
\name{fasta_to_line}
\alias{fasta_to_line}
\title{Converts fasta file to line}
\usage{
fasta_to_line(fasta, path_db = NULL)
}
\arguments{
\item{fasta}{A fasta file or the path to one.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe with the collumns "k-mer_id", "sequence", organized like a line file
}
\description{
Converts fasta file to line
}
\examples{
# 1. Create temporary fasta file
fasta_temp <- tempfile(fileext = ".fasta")
writeLines(
text = c(
">seq1", "ATCGATCG",
">seq2", "TCGATCGA"),
con = fasta_temp
)
# 2. Run function with temporary file
line_file <- fasta_to_line(fasta_temp)
# 3. View resulting line dataframe
print(line_file)
# 4. Delete temporary file
unlink(fasta_temp)
}
