% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters_1_seq.R
\name{get_clusters_1_seq}
\alias{get_clusters_1_seq}
\title{Selects the clusters that contain only one sequence}
\usage{
get_clusters_1_seq(dataframe_cd, path_db = NULL)
}
\arguments{
\item{dataframe_cd}{A dataframe from the run_cd_hit function}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe containing the collumns "cluster", "id", "tag".
}
\description{
Selects the clusters that contain only one sequence
}
\examples{
# 1. Create temporary dataframe
cdhit_out <- data.frame(V1 = c(1, 1, 2, 3, 3),
 V2 = c("A_1", "B_13", "C_69", "A_4", "A_2"))

# 2. Run function with  dataframe
cluster_1_seq <- get_clusters_1_seq(cdhit_out)
# 3. View result
print(cluster_1_seq)
}
