% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exclusive_clusters.R
\name{get_exclusive_clusters}
\alias{get_exclusive_clusters}
\title{Selects the clusters that contain only one marker}
\usage{
get_exclusive_clusters(dataframe_cd, path_db = NULL)
}
\arguments{
\item{dataframe_cd}{A dataframe from the run_cd_hit function}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe containing the collumns "cluster", "id", "tag".
}
\description{
Selects the clusters that contain only one marker
}
\examples{
# 1. Create temporary dataframe
cdhit_out <- data.frame(V1 = c(1, 1, 2, 3, 3),
 V2 = c("A_1", "B_13", "C_69", "A_4", "A_2"))

# 2. Run function with  dataframe
tag_exclusive_clusters <- get_exclusive_clusters(cdhit_out)
# 3. View result
print(tag_exclusive_clusters)
}
