% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_sharing_relation.R
\name{kmer_sharing_relation}
\alias{kmer_sharing_relation}
\title{Creates a Venn Diagram based on the kmer sequence distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more markes it c}
\usage{
kmer_sharing_relation(..., path_db = NULL)
}
\arguments{
\item{...}{A list of at least two dataframes from containing renamed headers and sequences}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
A list of lists containing the the sequence and in which datasets the kmer is present
}
\description{
Creates a Venn Diagram based on the kmer sequence distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more markes it c
}
\examples{
# 1. Create temporary dataframes
data_1 <- data.frame(id = c("tag1_1_1", "tag1_2_2", "tag1_3_3", "tag1_4_4", "tag1_5_5"),
 seq = c("AAAG", "AGAG", "CTGC", "AAAA", "GAGA"))
data_2 <- data.frame(id = c("tag2_1_1", "tag2_2_2", "tag2_3_3", "tag2_4_4", "tag2_5_5"),
 seq = c("AATG", "AGAG", "CAGC", "AATA", "GAGA"))

# 2. Run function with temporary file
dataset_sharing_relation <- kmer_sharing_relation(data_1, data_2)
# 3. View result
print(dataset_sharing_relation)
}
