% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clstr_file.R
\name{load_clstr_file}
\alias{load_clstr_file}
\title{Load clstr file from CD-hit in environment}
\usage{
load_clstr_file(file)
}
\arguments{
\item{file}{A clstr file from CD-hit.}
}
\value{
A dataframe containing the clustered groups
}
\description{
Load clstr file from CD-hit in environment
}
\examples{
# 1. Load clstr file path
clstr_path <- system.file("extdata",
                          "git_hub_test",
                          "clustered_datasets.clstr",
                           package = "AEkcompgen")
# 2. Run load_clstr_file
clstr_df <- load_clstr_file(clstr_path)
# 3. View resulting line dataframe
print(clstr_df)

}
