% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fasta.R
\name{load_fasta}
\alias{load_fasta}
\title{Converts fasta file to line}
\usage{
load_fasta(fasta_file)
}
\arguments{
\item{fasta_file}{A fasta file or the path to one.}
}
\value{
A dataframe with the collumns "k-mer_id", "sequence", organized like a line file
}
\description{
Converts fasta file to line
}
\examples{
# 1. Create temporary fasta file
fasta_temp <- tempfile(fileext = ".fasta")
writeLines(
text = c(
">seq1", "ATCGATCG",
">seq2", "TCGATCGA"),
con = fasta_temp
)
# 2. Run function with temporary file
fasta <- load_fasta(fasta_temp)
# 3. View resulting line dataframe
print(fasta)
# 4. Delete temporary file
unlink(fasta_temp)
}
