% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fasta_jf.R
\name{load_fasta_jf}
\alias{load_fasta_jf}
\title{Loads jellyfish dump fasta output into a dataframe, and marks its sequence by renaming the header}
\usage{
load_fasta_jf(fasta_file, marker)
}
\arguments{
\item{fasta_file}{A fasta output from jellyfish dump contaning the kmer count in the header}

\item{marker}{A marker to be added to the header of the fasta, identifying it}
}
\value{
A dataframe containing the counted kmers and their headers with the frequency information
}
\description{
Loads jellyfish dump fasta output into a dataframe, and marks its sequence by renaming the header
}
\examples{
# 1. Create temporary fasta file
fasta_temp <- tempfile(fileext = ".fasta")
writeLines(
text = c(">10", "ATCGATCG",
">13", "TCGATCGA"),
con = fasta_temp
)
# 2. Run function with temporary file
line_file <- load_fasta_jf(fasta_file = fasta_temp, marker = "tag")
# 3. View resulting line dataframe
print(line_file)
# 4. Delete temporary file
unlink(fasta_temp)
}
