% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped_file_to_fasta.R
\name{mapped_file_to_fasta}
\alias{mapped_file_to_fasta}
\title{Turns bam/sam file into a fasta file.}
\usage{
mapped_file_to_fasta(input, output)
}
\arguments{
\item{input}{A sorted bam/sam file, (run_bwa already sorts its output)}

\item{output}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
Saves in the current directory a fasta file containing the reads information from sam/bam file.
}
\description{
Turns bam/sam file into a fasta file.
}
\examples{
\dontshow{if (Sys.which("samtools") != "") withAutoprint(\{ # examplesIf}
#1. Load .bam test file
temp_bam <- system.file("extdata", "test_mapped.bam", package = "AEkcompgen")
#2. Run mapped_file_to_fasta function
mapped_file_to_fasta(input =  temp_bam, output = "temp_bam")
#3. Remove temporary file
unlink("temp_bam.fasta")
\dontshow{\}) # examplesIf}
}
