% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_jellyfish_histo.R
\name{mark_jellyfish_histo}
\alias{mark_jellyfish_histo}
\title{Adds a marker collumn to a dataframe, or loaded file of a histogram}
\usage{
mark_jellyfish_histo(histo_file = NULL, histo_df = NULL, marker, out = NULL)
}
\arguments{
\item{histo_file}{A histogram file from jellyfish histo}

\item{histo_df}{A loaded dataframe from the run_jellyfish histo function}

\item{marker}{The marker tag that will be added to the dataframe}

\item{out}{The name of the file in which the histogram will be saved}
}
\value{
A marked histogram containing the k-mer frequency and count of the k-mers in the .jf file
}
\description{
Adds a marker collumn to a dataframe, or loaded file of a histogram
}
\examples{
# 1. Create temporary dataframes
temp_file <- tempfile(fileext = ".histo")
writeLines(
text = c(
"1 100",
"2 150",
"3 50"),
con = temp_file
)

# 2. Run function with temporary file
marked_histo <- mark_jellyfish_histo(histo_file = temp_file, marker = "temp")
# 3. View result
print(marked_histo)
unlink(temp_file)
}
