% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_euler.R
\name{markers_euler}
\alias{markers_euler}
\title{Creates an Euler Diagram based on the cluster distribuition among markers
The suggested maximum number of markers is 4.}
\usage{
markers_euler(collapsed_markers, marker_label = NULL, color_pallete = "Set 2")
}
\arguments{
\item{collapsed_markers}{A dataframe from the collapse_markers function}

\item{marker_label}{A list with the names of the markers wished to be in the diagram, KEEP THE MARKERS' ORDER IN MIND}

\item{color_pallete}{A color pallete native from R, default is "Set 2"}
}
\value{
A Euler diagram based on the cluster distribuition among markers
}
\description{
Creates an Euler Diagram based on the cluster distribuition among markers
The suggested maximum number of markers is 4.
}
\examples{
# 1. Create temporary dataframe
collapsed_markers_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))

# 2. Run function with temporary file
euler_cd <- markers_euler(collapsed_markers_test)
# 3. View result
print(euler_cd)
}
