% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_euler_kmers.R
\name{markers_euler_kmers}
\alias{markers_euler_kmers}
\title{Creates an Euler Diagram based on the kmer sequence distribuition among markers.
The suggested maximum number of markers is 4 but can hold it up to 7 markers, if there are more markes it c}
\usage{
markers_euler_kmers(
  ...,
  path_db = NULL,
  marker_label = NULL,
  color_pallete = "PuBuGn"
)
}
\arguments{
\item{...}{A list of at least two dataframes from containing renamed headers and sequences}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}

\item{marker_label}{A list with the names of the markers wished to be in the diagram, KEEP THE MARKERS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}
}
\value{
An Euler Diagram based on the kmer distribuition among markers
}
\description{
Creates an Euler Diagram based on the kmer sequence distribuition among markers.
The suggested maximum number of markers is 4 but can hold it up to 7 markers, if there are more markes it c
}
\examples{
# 1. Create temporary dataframes
data_1 <- data.frame(id = c("tag1_1_1", "tag1_2_2", "tag1_3_3", "tag1_4_4", "tag1_5_5"),
 seq = c("AAAG", "AGAG", "CTGC", "AAAA", "GAGA"))
data_2 <- data.frame(id = c("tag2_1_1", "tag2_2_2", "tag2_3_3", "tag2_4_4", "tag2_5_5"),
 seq = c("AATG", "AGAG", "CAGC", "AATA", "GAGA"))

# 2. Run function with temporary file
venn_k <- markers_euler_kmers(data_1, data_2)
# 3. View result
print(venn_k)
}
