% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_upset_clusters.R
\name{markers_upset_clusters}
\alias{markers_upset_clusters}
\title{Creates an UpSet Plot based on the cluster distribuition among markers.
Recommeded whe working with a lot of markers}
\usage{
markers_upset_clusters(
  collapsed_markers,
  marker_label = NULL,
  color_pallete = "PuBuGn",
  empty_intersection = NULL,
  file_name = NULL
)
}
\arguments{
\item{collapsed_markers}{A dataframe from the collapse_markers function}

\item{marker_label}{A list with the names of the markers wished to be in the diagram, KEEP THE MARKERS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}

\item{empty_intersection}{If the empty intersectios are to be displayed put "on" or not, default is NULL}

\item{file_name}{The name for the file in case you want the plot to be saved automatically}
}
\value{
A Venn Diagram based on the cluster distribuition among markers
}
\description{
Creates an UpSet Plot based on the cluster distribuition among markers.
Recommeded whe working with a lot of markers
}
\examples{
# 1. Create temporary dataframe
collapsed_markers_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))

# 2. Run function with temporary file
upset_cd <- markers_upset_clusters(collapsed_markers_test)
# 3. View result
print(upset_cd)
}
