% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_venn.R
\name{markers_venn}
\alias{markers_venn}
\title{Creates a Venn Diagram based on the cluster distribuition among markers.
The suggested maximum number of markers is 4 but can hold it up to 7 markers, if there are more markes it c}
\usage{
markers_venn(collapsed_markers, marker_label = NULL, color_pallete = "PuBuGn")
}
\arguments{
\item{collapsed_markers}{A dataframe from the collapse_markers function}

\item{marker_label}{A list with the names of the markers wished to be in the diagram, KEEP THE MARKERS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}
}
\value{
A Venn Diagram based on the cluster distribuition among markers
}
\description{
Creates a Venn Diagram based on the cluster distribuition among markers.
The suggested maximum number of markers is 4 but can hold it up to 7 markers, if there are more markes it c
}
\examples{
# 1. Create temporary dataframe
collapsed_markers_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))

# 2. Run function with temporary file
venn_cd <- markers_venn(collapsed_markers_test)
# 3. View result
print(venn_cd)
}
