% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_1_seq.R
\name{plot_cluster_1_seq}
\alias{plot_cluster_1_seq}
\title{Plots the quantity of clusters with 1 k-mer, and the kmer count}
\usage{
plot_cluster_1_seq(
  cluster_1_seq_df,
  dataset_label = NULL,
  free_scale = FALSE,
  color_palette = "Set2",
  plot_title = "Not clustered k-mer distribution"
)
}
\arguments{
\item{cluster_1_seq_df}{Dataframe from  the get_clusters_1_seq function.}

\item{dataset_label}{A list with the names of the labels wished to be in the diagram, KEEP THE labelS' ORDER IN MIND}

\item{free_scale}{If the plots by labels should be free scale, put TRUE, or in the same scale, default.}

\item{color_palette}{A color palette from the scale_fill_gradientn function, default is "Set2".}

\item{plot_title}{Change the title of the plot.}
}
\value{
A Venn Diagram based on the cluster distribuition among labels
}
\description{
Plots the quantity of clusters with 1 k-mer, and the kmer count
}
\examples{
# 1. Create temporary dataframe
cluster_1_seq_df <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A_1_69", "B_1_34", "C_1_35", "B_2_8", "C_2_9"),
  V3 = c("A", "B", "C", "B", "C"))

# 2. Run function with temporary file
plots_1_seq <- plot_cluster_1_seq(cluster_1_seq_df)
# 3. View result
print(plots_1_seq)
}
