% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_euler_clusters.R
\name{plot_euler_clusters}
\alias{plot_euler_clusters}
\title{Creates a Euler Diagram based on the cluster distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more datasets the diagram may not make sense}
\usage{
plot_euler_clusters(
  cluster_sharing_relation,
  dataset_label = NULL,
  color_pallete = "PuBuGn",
  plot_title = "Cluster distribution across datasets"
)
}
\arguments{
\item{cluster_sharing_relation}{A list from the cluster_sharing_relation function}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE datasetS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}

\item{plot_title}{The title for the plot.}
}
\value{
An Euler Diagram based on the cluster distribuition among datasets
}
\description{
Creates a Euler Diagram based on the cluster distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more datasets the diagram may not make sense
}
\examples{
# 1. Create temporary dataframe
collapsed_datasets_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))
# 2. Run cluster_sharing_relation function to get list
dataset_sharing_relation <- cluster_sharing_relation(collapsed_datasets_test)
# 3. Run function with temporary file
euler_cd <- plot_euler_clusters(dataset_sharing_relation)
# 3. View result
print(euler_cd)
}
