% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram_kmer_freq.R
\name{plot_histogram_kmer_freq}
\alias{plot_histogram_kmer_freq}
\title{Plots an histogram based on the k-mer count and frequency}
\usage{
plot_histogram_kmer_freq(..., normalization_value = NULL, colors = NULL)
}
\arguments{
\item{...}{Dataframes from the mark_jellyfish histo function}

\item{normalization_value}{A value to divide the frequency of k-mers, usually based on the coverage of the mapped reads.}

\item{colors}{The color or list of colors to fill the plot's bars.}
}
\value{
A plot objetct displaying a k-mer frequency hostogram:
\item{X-axis}{K-mer frequency} the number of times a k-mer was counted.
\item{Y-axis}{Count} Quantity of distinct k-mers that appeared the number of times indicated in the x axis.
}
\description{
Plots an histogram based on the k-mer count and frequency
}
\examples{
# 1. Create temporary dataframes
data_frame_test <- data.frame(id = c("1", "2", "3"),
seq = c("100", "150", "50"),
tag = c("1", "2", "2"))
# 2. Run function with temporary file
histogram <- plot_histogram_kmer_freq(data_frame_test)
# 3. View result
print(histogram)
}
