% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset_clusters.R
\name{plot_upset_clusters}
\alias{plot_upset_clusters}
\title{Creates a Upset Plot based on the cluster distribuition among datasets.}
\usage{
plot_upset_clusters(
  cluster_sharing_relation,
  dataset_label = NULL,
  color_pallete = "PuBuGn",
  empty_intersection = NULL
)
}
\arguments{
\item{cluster_sharing_relation}{A list from the cluster_sharing_relation function}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE datasetS' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}

\item{empty_intersection}{If the empty intersectios are to be displayed put "on" or not, default is NULL}
}
\value{
An Upset Plot based on the cluster distribuition among datasets
}
\description{
Creates a Upset Plot based on the cluster distribuition among datasets.
}
\examples{
# 1. Create temporary dataframe
collapsed_datasets_test <- data.frame(cluster = c(1, 2, 3, 4, 5),
 V2 = c("A, B, C", "B, C", "B, A", "A, C", "B"))
# 2. Run cluster_sharing_relation function to get list
dataset_sharing_relation <- cluster_sharing_relation(collapsed_datasets_test)
# 3. Run function with temporary file
upset_cd <- plot_upset_clusters(dataset_sharing_relation)
# 3. View result
print(upset_cd)
}
