% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn_kmers.R
\name{plot_venn_kmers}
\alias{plot_venn_kmers}
\title{Creates a Venn Diagram based on the kmer sequence distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more datasets the diagram may not make sense}
\usage{
plot_venn_kmers(
  kmer_sharing_relation,
  dataset_label = NULL,
  color_pallete = "PuBuGn",
  plot_title = "K-mer distribution across datasets"
)
}
\arguments{
\item{kmer_sharing_relation}{A list from the kmer_sharing_relation function}

\item{dataset_label}{A list with the names of the datasets wished to be in the diagram, KEEP THE datasets' ORDER IN MIND}

\item{color_pallete}{A color pallete from the scale_fill_gradientn function, default is "PuBuGn"}

\item{plot_title}{The title for the plot.}
}
\value{
A Venn Diagram based on the kmer distribuition among datasets
}
\description{
Creates a Venn Diagram based on the kmer sequence distribuition among datasets.
The suggested maximum number of datasets is 4 but can hold it up to 7 datasets, if there are more datasets the diagram may not make sense
}
\examples{
# 1. Create temporary dataframes
data_1 <- data.frame(id = c("tag1_1_1", "tag1_2_2", "tag1_3_3", "tag1_4_4", "tag1_5_5"),
 seq = c("AAAG", "AGAG", "CTGC", "AAAA", "GAGA"))
data_2 <- data.frame(id = c("tag2_1_1", "tag2_2_2", "tag2_3_3", "tag2_4_4", "tag2_5_5"),
 seq = c("AATG", "AGAG", "CAGC", "AATA", "GAGA"))
# 2. Run kmer_sharing_relation function to get list
dataset_sharing_relation <- kmer_sharing_relation(data_1, data_2)
# 3. Run function with temporary file
venn_k <- plot_venn_kmers(dataset_sharing_relation)
# 4. View result
print(venn_k)
}
