% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_sequence.R
\name{retrieve_sequence}
\alias{retrieve_sequence}
\title{Retrieves the sequence information from a dataframe containing only the id and cluster information.}
\usage{
retrieve_sequence(clustered_df, seq_df, fasta_file_name = NULL, path_db = NULL)
}
\arguments{
\item{clustered_df}{A dataframe containing the Cluster and id, k-mer identification, collumns.}

\item{seq_df}{A dataframe containing the seq, sequence of the k-mer, and id, kmer identification, collumns.}

\item{fasta_file_name}{If want to save a fasta file containing the k-mers and their cluster info, give the name of the file.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
A dataframe containing the Cluster, id and seq information.
}
\description{
Retrieves the sequence information from a dataframe containing only the id and cluster information.
}
\examples{
# 1. Create temporary fasta file
data_frame_cluster <- data.frame(cluster = c(1, 1, 2),
 id = c("dataset1_1_28", "dataset2_1_33", "dataset1_2_2"))
data_frame_seq <- data.frame(id = c("dataset1_1_28", "dataset2_1_33", "dataset1_2_2"),
seq = c("GACAGGTACAAGAAGGAGTA", "AGGGCGACCTTCGATTCGGA", "TTTACACACTCTCCTTGGAC"))
# 2. Run function with temporary file
seq_info_cluster_df <- retrieve_sequence(data_frame_cluster, data_frame_seq)
# 3. View resulting line dataframe
print(seq_info_cluster_df)
}
