% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bwa.R
\name{run_bwa}
\alias{run_bwa}
\title{Runs bwa mem to map the given sequencing files into a reference genome.
If given the variable, also runs bwa mem and samtools to remove possible contaminated reads from a contaminant_ref reference file.
Also if given another variable, runs samtools to verify mapping quality and filter reads based on it.
Works with short reads from the Illumina sequencing platform.}
\usage{
run_bwa(
  input,
  output,
  ref,
  contaminant_ref = NULL,
  mapping_quality = NULL,
  output_format = "bam",
  keep_index = TRUE
)
}
\arguments{
\item{input}{Input FASTQ files/paths.}

\item{output}{Mapped reads in the given reference genome.}

\item{ref}{Reference genome where the reads will be mapped.}

\item{contaminant_ref}{Optional: Reference contaminant_ref genome, where reads that have matched with will be removed.}

\item{mapping_quality}{Optional: Value from 0 to 60, to filter reads with mapping quality equal or above the given value.}

\item{output_format}{Format for the output of the mapping file (bam/sam), default is "bam' for binary file.}

\item{keep_index}{Optional: Default is true, put false if you dont want to keep the index files in the directory for faster performance.}
}
\description{
Runs bwa mem to map the given sequencing files into a reference genome.
If given the variable, also runs bwa mem and samtools to remove possible contaminated reads from a contaminant_ref reference file.
Also if given another variable, runs samtools to verify mapping quality and filter reads based on it.
Works with short reads from the Illumina sequencing platform.
}
\examples{
\dontshow{if (Sys.which("bwa") != "" && Sys.which("samtools") != "") withAutoprint(\{ # examplesIf}
#1. Load .fastq test file
temp_fastq <- system.file("extdata", "test.fastq", package = "AEkcompgen")
temp_ref_fasta <- system.file("extdata", "test_reference.fasta", package = "AEkcompgen")
#2. Run run_bwa function
run_bwa(input = temp_fastq,
 output = "test_mapped",
 ref = temp_ref_fasta,
 keep_index = FALSE)
#3. Unlink ouput file
unlink("test_mapped.bam")
\dontshow{\}) # examplesIf}
}
