% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cd_hit.R
\name{run_cd_hit}
\alias{run_cd_hit}
\title{CD-HIT-est for the given dataframe or fasta_file}
\usage{
run_cd_hit(data_frame = NULL, fasta_file = NULL, identity, output = NULL)
}
\arguments{
\item{data_frame}{A dataframe containing id, seq and marker columns.}

\item{fasta_file}{A fasta file or a path to one.}

\item{identity}{The identity degree of similarity that will be used to cluster the sequences.}

\item{output}{If you want to save the output file, the give the name for it.}
}
\value{
A dataframe containing the counted kmers and their headers with the frequency information
}
\description{
CD-HIT-est for the given dataframe or fasta_file
}
\examples{
# 1. Create temporary dataframe
data_frame_test <- data.frame(id = c("seq1", "seq2", "seq3"),
seq = c("GACAGGTACAAGAAGGAGTA", "AGGGCGACCTTCGATTCGGA", "TTTACACACTCTCCTTGGAC")
)
# 2. Run function with temporary file
cd_hit_test <- run_cd_hit(data_frame = data_frame_test, identity = 0.9)
# 3. View resulting line dataframe
print(cd_hit_test)

}
