% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Saves plots.
In pdf or png format.}
\usage{
save_plot(plot, file_name, format = "pdf", width = NULL, height = NULL)
}
\arguments{
\item{plot}{A plot.}

\item{file_name}{Name to save the plot.}

\item{format}{Format to save the file, between "pdf" and "png".}

\item{width}{Width of the pdf or of the png file}

\item{height}{Height of the pdf or of the png file}
}
\value{
A Euler diagram based on the cluster distribuition among markers
}
\description{
Saves plots.
In pdf or png format.
}
\examples{
# 1. Create temporary plot
plot_test <- plot(1, type = "n",
    xlim = c(0, 10), ylim = c(0, 10),
    xlab = "X Axis Test", ylab = "Y Axis Test",
    main = "Empty Test Plot")
# 2. Run function with temporary file
save_plot(plot_test, file_name = "plot_test", format = "pdf")
# 4. Delete temporary file
unlink("plot_test.pdf")
}
