% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_marker.R
\name{set_marker}
\alias{set_marker}
\title{Add a marker to a fasta file}
\usage{
set_marker(fasta, marker, path_db = NULL)
}
\arguments{
\item{fasta}{A fasta file or the path to one.}

\item{marker}{The tag term that will tag the sequences in this file.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe with the collumns "k-mer_id", "sequence" and "tag".
}
\description{
Add a marker to a fasta file
}
\examples{
marked_file <- set_marker("../fasta.fasta.txt", "Leish")
}
