% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_marker_df.R
\name{set_marker_df}
\alias{set_marker_df}
\title{Add a marker to a dataframe}
\usage{
set_marker_df(df, marker)
}
\arguments{
\item{df}{A dataframe.}

\item{marker}{The tag term that will tag the sequences in this file.}
}
\value{
A dataframe with the collumns "k-mer_id", "sequence" and "tag".
}
\description{
Add a marker to a dataframe
}
\examples{
# 1. Create temporary dataframe
dataframe_test <- data.frame(id = c("seq_1", "seq_2"),
 seq = c("ATCGATCG", "TCGATCGA"))
# 2. Run function with temporary file
marked_file <- set_marker_df(dataframe_test, "Leish")
# 3. View result
print(marked_file)
}
