% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_marker_cd_hit.R
\name{set_markers_cd_hit}
\alias{set_markers_cd_hit}
\title{Add a marker to a fasta file}
\usage{
set_markers_cd_hit(input, markers = character(), path_db = NULL)
}
\arguments{
\item{input}{A file from cd_hit or formated like one.}

\item{markers}{A list of tag term that will tag the clusters within this file.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe containing the collumns "cluster", "id", "tag".
}
\description{
Add a marker to a fasta file
}
\examples{
# 1. Create temporary fasta file
cdhit_temp <- tempfile(fileext = ".txt")
writeLines(
text = c(
"1  infantum_1",
"1  major_1",
"2  infantum_2",
"3  infantum_3",
"3  major_2"),
con = cdhit_temp
)
# 2. Run function with temporary file
marked_file <- set_markers_cd_hit(cdhit_temp, c("infantum", "major"))
# 3. View result
print(marked_file)
# 4. Delete temporary file
unlink(cdhit_temp)
}
