% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_sequencing.R
\name{subsample_sequencing}
\alias{subsample_sequencing}
\title{Subsample the given input bam/sam file to the wanted average sequencing depth.}
\usage{
subsample_sequencing(
  input,
  current_average_depth,
  wanted_average_depth,
  output,
  output_format = "bam",
  path_db = NULL
)
}
\arguments{
\item{input}{A sorted bam/sam file, (run_bwa already sorts its output)}

\item{current_average_depth}{The current average sequencing depth of the input file.}

\item{wanted_average_depth}{The wanted average sequencing depth.}

\item{output}{The path for the subsampled bam/sam file.}

\item{output_format}{Format for the output of the mapping file (bam/sam), default is "bam' for binary file.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
Saves the subsampled file in the output given path, and prints the current average depth calculated
}
\description{
Subsample the given input bam/sam file to the wanted average sequencing depth.
}
\examples{
\dontshow{if (Sys.which("samtools") != "") withAutoprint(\{ # examplesIf}
#1. Load .bam test file
temp_bam <- system.file("extdata", "test_mapped.bam", package = "AEkcompgen")
#2. Run subsample_sequencing function
subsample_sequencing(input =  temp_bam,
 current_average_depth = 1,
 wanted_average_depth = 0.8,
 output = "test_mapped_subsampled",
 output_format = "bam")
#3. Delete file
unlink("test_mapped_subsampled.bam")
\dontshow{\}) # examplesIf}
}
