% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_hit_info.R
\name{cd_hit_info}
\alias{cd_hit_info}
\title{Gives info of the run_cd_hit output.
Such as quantity of clusters, mean, median, largest cluster and smallest cluster.}
\usage{
cd_hit_info(dataframe_cd, path_db = NULL, out = NULL)
}
\arguments{
\item{dataframe_cd}{A dataframe from the run_cd_hit function}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}

\item{out}{The name or path of a file to save the information on a .txt file}
}
\value{
A dataframe containing the collumns "cluster", "id", "tag".
}
\description{
Gives info of the run_cd_hit output.
Such as quantity of clusters, mean, median, largest cluster and smallest cluster.
}
\examples{
# 1. Create temporary dataframe
cdhit_out <- data.frame(V1 = c(1, 1, 2, 3, 3),
 V2 = c("infantum_1_28", "major_1_33", "infantum_2_2", "infantum_3_19", "major_2_49"))

# 2. Run function with dataframe
marked_file <- cd_hit_info(cdhit_out)
# 3. View result
print(marked_file)
}
