% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_diff_shared_kmers.R
\name{plot_count_diff_shared_kmers}
\alias{plot_count_diff_shared_kmers}
\title{Compares the kmer frequency between different possible pairs of dataset labels,
Takes output of the select_shared_seq and compares the kmer frequency between two labels giving the modulus
of the difference.}
\usage{
plot_count_diff_shared_kmers(
  shared_df,
  cut_off = NULL,
  color_palette = "Set2",
  free_scale = FALSE,
  plot_title = "Modulus of Count Difference"
)
}
\arguments{
\item{shared_df}{Dataframe from select_shared_seq.}

\item{cut_off}{A minimal value of k-mer count difference.}

\item{color_palette}{A color pallete from the scale_fill_gradientn function, default is "Set2".}

\item{free_scale}{If the plots by labels should be free scale, put TRUE, or in the same scale, default.}

\item{plot_title}{The title for the plot.}
}
\value{
A daframe with the frequency modulus difference of different pairs of dataset labels.
}
\description{
Compares the kmer frequency between different possible pairs of dataset labels,
Takes output of the select_shared_seq and compares the kmer frequency between two labels giving the modulus
of the difference.
}
\examples{
# 1. Create temporary dataframe
shared_test <- data.frame(seq = c("GAGA", "ATGA", "AAAA"),
 label = c("label1,label2", "label1,label2", "label1,label2"),
  id_1 = c("label1_3_3", "label1_2_15", "label1_3_25"),
  id_2 = c("label2_3_45", "label2_2_50",
   "label2_3_30"))

# 2. Run function with temporary file
plots_freq_comp <- plot_count_diff_shared_kmers(shared_test)
# 3. View result
print(plots_freq_comp)
}
