% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_diff_shared_kmers.R
\name{plot_count_diff_shared_kmers}
\alias{plot_count_diff_shared_kmers}
\title{Compares the kmer frequency between two markers
Takes output of the select_shared_seq and compares the kmer frequency between two markers giving the modulus
of the difference.}
\usage{
plot_count_diff_shared_kmers(
  shared_df,
  cut_off = NULL,
  separate_markers = FALSE,
  colors = c("purple", "green"),
  path_db = NULL
)
}
\arguments{
\item{shared_df}{Dataframe from select_shared_seq.}

\item{cut_off}{A minimal value of kmer_frequency}

\item{separate_markers}{Separates markers into different plots, using facet_wrap in ggplot.}

\item{colors}{A list of two colors to fill the plot's bars.}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file.}
}
\value{
A daframe with the frequency modulus difference.
}
\description{
Compares the kmer frequency between two markers
Takes output of the select_shared_seq and compares the kmer frequency between two markers giving the modulus
of the difference.
}
\examples{
# 1. Create temporary dataframe
shared_test <- data.frame(seq = c("GAGA", "ATGA", "AAAA"),
 marker = c("Marker1,Marker2", "Marker1,Marker2", "Marker1,Marker2"),
  header_1 = c("Marker1_3_3", "Marker1_2_15", "Marker1_3_25"),
  header_2 = c("Marker2_3_45", "Marker2_2_50", "Marker2_3_30"))

# 2. Run function with temporary file
plots_freq_comp <- plot_count_diff_shared_kmers(shared_test)
# 3. View result
print(plots_freq_comp)
}
