% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fasta_jf.R
\name{rename_fasta_jf}
\alias{rename_fasta_jf}
\title{Loads jellyfish dump fasta output into a dataframe, and renames the header by adding the dataset_label and the rownumber}
\usage{
rename_fasta_jf(fasta_file, dataset_label, out = NULL)
}
\arguments{
\item{fasta_file}{A fasta output from jellyfish dump contaning the kmer count in the header}

\item{dataset_label}{A label to be added to the header of the fasta, identifying it}

\item{out}{The name of a path to save the renamed fasta file}
}
\value{
A dataframe containing the counted kmers and their headers with the frequency information
}
\description{
Loads jellyfish dump fasta output into a dataframe, and renames the header by adding the dataset_label and the rownumber
}
\examples{
# 1. Create temporary fasta file
fasta_temp <- tempfile(fileext = ".fasta")
writeLines(
text = c(">10", "ATCGATCG",
">13", "TCGATCGA"),
con = fasta_temp
)
# 2. Run function with temporary file
line_file <- rename_fasta_jf(fasta_file = fasta_temp, dataset_label = "tag")
# 3. View resulting line dataframe
print(line_file)
# 4. Delete temporary file
unlink(fasta_temp)
}
