% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bwa_remove_contaminant.R
\name{run_bwa_remove_contaminant}
\alias{run_bwa_remove_contaminant}
\title{Runs bwa mem to remove contaminant based on what was mapped in the contaminant reference gennome.
Works with short reads from the Illumina sequencing platform.}
\usage{
run_bwa_remove_contaminant(input, output, ref, keep_index = TRUE)
}
\arguments{
\item{input}{Input FASTQ files/paths.}

\item{output}{FASTQ file with the reads that mapped in the contaminant reference genome removed.}

\item{ref}{Reference contaminant genome where the reads will be mapped.}

\item{keep_index}{Optional: Default is true, put false if you dont want to keep the index files in the directory for faster performance.}
}
\description{
Runs bwa mem to remove contaminant based on what was mapped in the contaminant reference gennome.
Works with short reads from the Illumina sequencing platform.
}
\examples{
\dontshow{if (Sys.which("bwa") != "" && Sys.which("samtools") != "") withAutoprint(\{ # examplesIf}
#1. Load .fastq test file
temp_fastq <- system.file("extdata", "test.fastq", package = "AEkcompgen")
temp_ref_fasta <- system.file("extdata", "test_reference.fasta", package = "AEkcompgen")
#2. Run run_bwa_remove_contaminant function
run_bwa_remove_contaminant(input = temp_fastq,
 output = "test_mapped",
 ref = temp_ref_fasta,
 keep_index = FALSE)
#3. Unlink output file
unlink("test_mapped.fastq.gz")
\dontshow{\}) # examplesIf}
}
