% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jellyfish.R
\name{run_jellyfish}
\alias{run_jellyfish}
\title{Runs jellyfish and give it's output, both in dataframe form and saving it in the directory given in the output argument
The id for the sequence contains the row of the kmer in the fasta file and the the count of the k-mer}
\usage{
run_jellyfish(
  data_frame = NULL,
  fasta_file = NULL,
  length,
  marker,
  lower_count = NULL,
  upper_count = NULL,
  out
)
}
\arguments{
\item{data_frame}{A dataframe containing the ID and sequence collumns}

\item{fasta_file}{A fasta file or the path to one.}

\item{length}{The length of kmers that will be counted}

\item{marker}{The marker that will be used in the header of sequence and file names}

\item{lower_count}{Don't output k-mers with count lower than this}

\item{upper_count}{Don't output k-mers with count higher than this}

\item{out}{Name for output fasta file}
}
\value{
A dataframe containing the counted kmers and their headers with the frequency information
}
\description{
Runs jellyfish and give it's output, both in dataframe form and saving it in the directory given in the output argument
The id for the sequence contains the row of the kmer in the fasta file and the the count of the k-mer
}
\examples{
\dontshow{if (Sys.which("jellyfish") != "") withAutoprint(\{ # examplesIf}
# 1. Create temporary fasta file
data_frame_test <- data.frame(id = c("seq1", "seq2", "seq3"),
seq = c("GACAGGTACAAGAAGGAGTA", "AGGGCGACCTTCGATTCGGA", "TTTACACACTCTCCTTGGAC"))
# 2. Run function with temporary file
line_file <- run_jellyfish(data_frame = data_frame_test,
 length =  10,
 marker = "test",
 out = "test_output")
# 3. View resulting line dataframe
print(line_file)
# 4. Delete temporary file
unlink("test_output.fasta")
unlink("temp.fasta")
unlink("test.jf")
unlink("temp_jellyfish_dump")
\dontshow{\}) # examplesIf}
}
