% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jellyfish.R
\name{run_jellyfish}
\alias{run_jellyfish}
\title{Runs jellyfish and gives it's output
The id for the sequence contains the row of the kmer in the fasta file and the the count of the k-mer}
\usage{
run_jellyfish(
  fasta_file,
  length,
  marker,
  lower_count = NULL,
  upper_count = NULL,
  histo = TRUE,
  path_db = NULL,
  output
)
}
\arguments{
\item{fasta_file}{A fasta file or the path to one.}

\item{length}{The length of kmers that will be counted}

\item{marker}{The marker that will be used in the header of sequence and file names}

\item{lower_count}{Don't output k-mers with count lower than this}

\item{upper_count}{Don't output k-mers with count higher than this}

\item{histo}{Make histogram based on kmer frequency}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}

\item{output}{Name for output fasta file}
}
\value{
A dataframe containing the counted kmers and their headers with the frequency information
}
\description{
Runs jellyfish and gives it's output
The id for the sequence contains the row of the kmer in the fasta file and the the count of the k-mer
}
\examples{
# 1. Create temporary fasta file
fasta_temp <- tempfile(fileext = ".fasta")
writeLines(
text = c(
">seq1", "GACAGGTACAAGAAGGAGTA",
">seq2", "AGGGCGACCTTCGATTCGGA",
">seq3", "TTTACACACTCTCCTTGGAC",
">seq4", "TGTGAACTTTTAAATTCGAT",
">seq5", "CACTTAAGGCTTGAAAACTA"),
con = fasta_temp
)
# 2. Run function with temporary file
line_file <- run_jellyfish(fasta_temp, length =  10, marker = "test", histo = FALSE, output = "test_output.fasta")
# 3. View resulting line dataframe
print(line_file)
# 4. Delete temporary file
unlink(fasta_temp)
unlink("test_output.fasta")
}
