% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jellyfish_histo.R
\name{run_jellyfish_histo}
\alias{run_jellyfish_histo}
\title{Runs the Jellyfish histo function}
\usage{
run_jellyfish_histo(jf_file, out = NULL)
}
\arguments{
\item{jf_file}{The .jf output file from jellyfish}

\item{out}{The name of the file in which the histogram will be saved}
}
\value{
A histogram containing the k-mer frequency and count of the k-mers in the .jf file
}
\description{
Runs the Jellyfish histo function
}
\examples{
\dontshow{if (Sys.which("jellyfish") != "") withAutoprint(\{ # examplesIf}
# 1. Load .jf test file
temp_jf <- system.file("extdata", "test.jf", package = "AEkcompgen")
# 2. Run function with temporary file
histo_file <- run_jellyfish_histo(temp_jf)
# 3. View result
print(histo_file)
\dontshow{\}) # examplesIf}
}
