% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trimmomatic.R
\name{run_trimmomatic}
\alias{run_trimmomatic}
\title{Runs trimmomatic to verify read quality using the quality score of phred 33, standar for Illumina data, trim and separate paired and unpaired reads.
Works with short reads from the Illumina sequencing platform.}
\usage{
run_trimmomatic(
  mode,
  input,
  out,
  minimun_read_length = 36,
  sliding_window_size = 4,
  average_phred = 15
)
}
\arguments{
\item{mode}{Which mode to run.
\describe{
\item{SE}{Single-end data, requires only one input FASTQ.}
\item{PE}{Paired-end data, requires two input FASTQ, one containing the forward and reverse sequences.}
}}

\item{input}{Input FASTQ files/paths.
\describe{
\item{1 file}{For single-end mode, give only one FASTQ file.}
\item{2 files}{If paired-end mode, give the forward and reverse sequen FASTQ.}
}}

\item{out}{Output FASTQ files names/paths(4 files if pair ended mode, 1 file for single ended mode)
\describe{
\item{1 output file}{If single-end mode, output contains only the reads that met the given phred/window_size requirements.}
\item{4 output files}{If paired-end mode, saving what was paired and unpaired, in this order, between forward and reverse sequences and met the given phred/window_size requirements.}
}}

\item{minimun_read_length}{Minimun read length to kept, default is 36.}

\item{sliding_window_size}{The sliding window size to verify average phred value, default is 4.}

\item{average_phred}{The average phred value for the sliding window, default is 15, if the phred value drops below it the read is cut.}
}
\description{
Runs trimmomatic to verify read quality using the quality score of phred 33, standar for Illumina data, trim and separate paired and unpaired reads.
Works with short reads from the Illumina sequencing platform.
}
\examples{
\dontshow{if (Sys.which("trimmomatic") != "") withAutoprint(\{ # examplesIf}
#1. Load .fastq test file
temp_fastq <- system.file("extdata", "test.fastq", package = "AEkcompgen")
#2. Run run_trimmomatic function
run_trimmomatic(mode = "SE", input = temp_fastq, out = "test_trimmed.fastq")
#3. Unlink ouput file
unlink("test_trimmed.fastq")
\dontshow{\}) # examplesIf}
}
