% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_exclusive_seq.R
\name{select_exclusive_seq}
\alias{select_exclusive_seq}
\title{Selects the exclusive sequences of two files after the set_marker() function}
\usage{
select_exclusive_seq(dataset_1, dataset_2, path_db = NULL)
}
\arguments{
\item{dataset_1}{Dataframe from set_marker() to be compared with dataset_2}

\item{dataset_2}{Dataframe from set_marker() to be compared with dataset_1}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe containing the exclusive sequences to each marker
}
\description{
Selects the exclusive sequences of two files after the set_marker() function
}
\examples{
# 1. Create temporary fasta file
fasta_temp1 <- tempfile(fileext = ".fasta")
writeLines(c(">seq1", "ATCGATCG",
">seq2", "TCGATCGA",
">seq3", "TCAGTCAG"),
con = fasta_temp1
)

fasta_temp2 <- tempfile(fileext = ".fasta")
writeLines(c(">seq1", "GCGCGCGC",
">seq2", "TATATATA",
">seq3", "TCAGTCAG"),
con = fasta_temp2
)
# 2. Run set_marker
marked_df_1 <- set_marker(fasta_temp1, "1")
marked_df_2 <- set_marker(fasta_temp2, "2")

# 3. Run function with temporary file
exc_seq <- select_exclusive_seq(marked_df_1, marked_df_2)
# 4. View result
print(exc_seq)
# 5. Delete temporary file
unlink(fasta_temp1)
unlink(fasta_temp2)
}
