% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_shared_seq.R
\name{select_shared_seq}
\alias{select_shared_seq}
\title{Selects the shared sequences of two files after the set_marker() function}
\usage{
select_shared_seq(..., path_db = NULL)
}
\arguments{
\item{...}{At least two dataframes containing the header and the seq column}

\item{path_db}{The path to save the .duckdb file, if none given it creates a temporary file}
}
\value{
A dataframe containing the exclusive sequences to each marker
}
\description{
Selects the shared sequences of two files after the set_marker() function
}
\examples{
# 1. Create temporary fasta file
fasta_temp1 <- data.frame(header = c("seq1", "seq2", "seq3", "seq4", "seq5"),
 seq = c("ATCGATCG", "TCGATCGA", "TCAGTCAG", "AATATATA", "TAGGGT"))

fasta_temp2 <- data.frame(header = c("seq6", "seq7", "seq8", "seq9", "seq10"),
 seq = c("GCGCGCG", "GAGAGAG", "TCTCTCTC", "AATATATA", "TAGGGT"))
# 3. Run function with example dataframes
exc_seq <- select_shared_seq(fasta_temp1, fasta_temp2)
# 4. View result
print(exc_seq)
}
